Fix describe features BF Cache result.
The code would left shift an 32 bit number so this code would fail to
report the correct descriptions for > 32 bit fields.

BUG=1192785

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2798051
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#868575}

+++ content/browser/renderer_host/back_forward_cache_can_store_document_result.cc
@@ -15,9 +15,9 @@ using blink::scheduler::WebSchedulerTrackedFeature;
 
 std::string DescribeFeatures(uint64_t blocklisted_features) {
   std::vector<std::string> features;
-  for (size_t i = 0;
-       i <= static_cast<size_t>(WebSchedulerTrackedFeature::kMaxValue); ++i) {
-    if (blocklisted_features & (1 << i)) {
+  for (uint32_t i = 0;
+       i <= static_cast<uint32_t>(WebSchedulerTrackedFeature::kMaxValue); ++i) {
+    if (blocklisted_features & (1ULL << i)) {
       features.push_back(blink::scheduler::FeatureToHumanReadableString(
           static_cast<WebSchedulerTrackedFeature>(i)));
     }
